validation error

+++ src/database/entities/umum/idea.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, HttpException, HttpStatus } from '@nestjs/common';
 import { Repository } from 'typeorm';
 import { InjectRepository } from '@nestjs/typeorm';
 
@@ -23,17 +23,29 @@ export class IdeaService {
   }
 
   async read(id: string) {
-    return await this.ideaRepository.findOne({ where: { id } });
+    const idea = await this.ideaRepository.findOne({ where: { id } });
+    if (!idea) {
+      throw new HttpException('Not Found', HttpStatus.NOT_FOUND);
+    }
+
+    return idea;
   }
 
   async update(id: string, data: Partial<IdeaDTO>) {
+    const idea = await this.ideaRepository.findOne({ where: { id } });
+    if (!idea) {
+      throw new HttpException('Not Found', HttpStatus.NOT_FOUND);
+    }
     await this.ideaRepository.update({ id }, data);
     return await this.ideaRepository.findOne({ id });
   }
 
   async destroy(id: string) {
+    const idea = await this.ideaRepository.findOne({ where: { id } });
+    if (!idea) {
+      throw new HttpException('Not Found', HttpStatus.NOT_FOUND);
+    }
     await this.ideaRepository.delete({ id });
-    const respons = await this.ideaRepository.find();
-    return { respons, deleted: 'deleted data succesfully' };
+    return { idea, deleted: 'deleted data succesfully' };
   }
 }
